// backend/src/types/code/userCode.ts
export enum UserErrorCode {
  EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
  CPF_ALREADY_EXISTS = "CPF_ALREADY_EXISTS",
  PHONE_ALREADY_EXISTS = "PHONE_ALREADY_EXISTS",
  USER_CREATE_FAILED = "USER_CREATE_FAILED",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  INVALID_PASSWORD = "INVALID_PASSWORD",
  LOGIN_FAILED = "LOGIN_FAILED",
  USER_FETCH_FAILED = "USER_FETCH_FAILED",
  USER_DELETE_FAILED = "USER_DELETE_FAILED",
  USER_UPDATE_FAILED = "USER_UPDATE_FAILED",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  SESSION_REVOKED = "SESSION_REVOKED",
  LOGOUT_FAILED = "LOGOUT_FAILED",
}